landscout/backend/requirements.txt
# LandScout Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.2
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# API Requests
httpx==0.25.1
requests==2.31.0

# Email
resend==0.7.0

# PDF Generation
reportlab==4.0.7

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
